replace data.table::shift function
just to be on the save side, the old code is still in the script to
see whether the changes really work [called nSx.o]. Please test!

Besides, Chrisoph mentions that the changed step is basically the
substraction of dead persons from still alive persons. In case it is
clear which variable represents what the function can be simplified
further.